+++ src/main/java/act/handler/SkipBuiltInEvents.java
@@ -22,10 +22,7 @@ package act.handler;
 
 import act.app.ActionContext;
 
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
+import java.lang.annotation.*;
 
 /**
  * Mark a request handling method shall skip result commit events triggering
@@ -34,6 +31,6 @@ import java.lang.annotation.Target;
  * {@link RequestHandler#skipEvents(ActionContext)} returns `true`.
  */
 @Retention(RetentionPolicy.RUNTIME)
-@Target(ElementType.METHOD)
+@Target({ElementType.METHOD, ElementType.TYPE})
 public @interface SkipBuiltInEvents {
 }